move container size selectors
